+++ src/index.js
@@ -104,10 +104,10 @@ export default class ReactMarkdownGithub extends Component {
   * @api private
   */
   transformLinkUri(uri, children, title) {
-    const { resolver } = this.props;
+    const { transformLinkUri } = this.props;
     const normalized = this.normalizeLinkUri(uri);
     const opts = { ...this.state, uri: normalized, children, title };
-    return resolver && resolver(opts) || normalized;
+    return transformLinkUri && transformLinkUri(opts) || normalized;
   }
 
   /**
@@ -170,7 +170,6 @@ export default class ReactMarkdownGithub extends Component {
   * @api private
   */
   render() {
-    const { className, source, resolver } = this.props;
     const renderers = {
       heading: this.renderHeading,
       ...this.props.renderers
@@ -178,16 +177,14 @@ export default class ReactMarkdownGithub extends Component {
 
     return (
       <ReactMarkdown
-        source={ source }
+        { ...this.props }
         renderers={ renderers }
-        className={ className }
-        resolver={ resolver }
         transformLinkUri={ this.transformLinkUri }
-        transformImageUri={ this.transformImageUri } />
+        transformImageUri={ this.transformImageUri }
+      />
     );
   }
 }
-
 
 ReactMarkdownGithub.propTypes = {
   /** {source} The Markdown content to be rendered by `ReactMarkdown` */
@@ -197,8 +194,8 @@ ReactMarkdownGithub.propTypes = {
    * e.g. https://github.mycorp.com/org/component/blob/master/README.md'
    */
   sourceUri: PropTypes.string,
-  /** {resolver} The callback function executed for each found URL */
-  resolver: PropTypes.func,
+  /** {transformLinkUri} The callback function executed for each found URL */
+  transformLinkUri: PropTypes.func,
   /** {transformImageUri} The callback function executed for each found image */
   transformImageUri: PropTypes.func,
   /** {renderers} the collection of resolvers to pass to `ReactMarkdown` */